Refactored the tests and resolvers
