+++ backend/src/create-datasources.ts
@@ -1,6 +1,6 @@
 import PostDatasource from "./datasources/post.datasource";
 import UserDatasource from "./datasources/user.datasource";
-import {users, posts} from "./seed-data";
+import { users, posts } from "./seed-data";
 
 const createDatasources = {
   postDatasource: new PostDatasource(posts),
